requirements.txt
streamlit
pandas
matplotlib
pyyaml>=6.0
openai>=1.50.0
